+++ lib/dynamic_library.dart
@@ -2,3 +2,4 @@ library dynamic_library;
 
 export 'src/wrapper_class.dart';
 export 'src/loader.dart';
+export 'src/frb_init.dart';